parser: Update IDENT name length limit rules
IDENTs are now limited to 25 characters instead of 38 to prevent
conflicts. This was made necessary by the prior commit to support
Testnet P2PKH addresses, which begin with an alphanumeric.

To make this less painful, this commit also lifts the limit entirely if
the IDENT name contains a `$`, `_` or `::`.

The CodeMirror syntax highlightning rules were updated accordingily,
with negative lookahead to tell addresses/xpubs/wifs apart from IDENTs.