psbt: Use the top-most Xpriv for signing, support multi-Xpriv
This is typically what the user wants, as the bip32_derivation and
tap_key_origins PSBT fields usually point to the fingerprint of the
top-most key and not to that of its child.

This also adds support for multi-Xprivs, which now became trivial as all
paths point to the same parent key.